backend/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pypdf
sentence-transformers
faiss-cpu
numpy
openai
